<fetch>: Добалвены тесты

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -74,6 +74,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void delete(Long id) {
+        if (id == null) {
+            throw new IncorrectParameterException("Id не может быть пустым!");
+        }
         userStorage.deleteById(id);
     }
